Replace use of elm-audio package with WebAudio
The HTML5 `Audio` object (used by elm-audio) introduces a perceptible
gap when looping. This issue goes away with use of the WebAudio API,
although it is more complex.

Currently, there is an (unpublished due to native code) elm-webaudio
package, but it doesn't appear fully updated for Elm 0.17, and also
includes way more functionality than needed by this app.

This commit implements the very basics of loading, playing, and
stopping a sound with WebAudio in Elm via a native JS module.