Handle jam overflows from `JamBalancer.refresh`
This makes `refresh` the single source of truth for jam capacity rather
than the removed `jam_capacity?`. This was necessary because of a race
condition that was present between asking the balancer if a particular
jam had capacity and tracking new users in the presence. The change
places new users in the balancer before they are tracked in the
presence.

Another key to preventing overflows was updating the user lists in the
balancer from the balancer state itself, rather than clobbering them
every time from the presence state. This was the source of another race
condition.

This commit has the nice side effect of simplifying the code in a bunch
of places.